requirements.txt
einops
safetensors
huggingface-hub
transformers
opencv-python
accelerate
datasets
tqdm
pillow
deepspeed
--extra-index-url https://download.pytorch.org/whl/cu126
torch
torchvision
torchaudio
